Optimize audio file for NyanCatMusicFormatter
Replace 8.7 MB MP3 file with a 433 KB version. The smaller file is also shorter
(27 seconds vs. 217 seconds), so I'm playing it in a loop.

This reduces the total gem size from 8.3 MB to 1.0 MB.

I've also rewritten the if osx? and if linux? statements as a single if...elsif
statement, since they should be mutually exclusive.