Alter how the music process is spawned
On the off chance that a user of the music formatter is using one of the
supported utilities somewhere else we really don't want to kill that for
them too. So I've reworked how music gets spawned to get us a PID so
when the music gets killed only the process the formatter spawned gets
killed. This also has the added benefit of being able to swap in an
arbitrary audio file supported by these utilities without having to
alter the length of time the code will wait before re-spawining the
music process.

Signed-off-by: Tyler Pickett <t.pickett66@gmail.com>